+++ src/Editor/Extension/ScriptAttributeUtilityExtension.cs
@@ -14,12 +14,20 @@ namespace CGame
         }
         
         private static MethodInfo _getDrawerTypeForPropertyAndTypeMethodInfo;
+        private static MethodInfo _getHandlerMethodInfo;
         
         public static Type GetDrawerTypeForPropertyAndType(SerializedProperty property, Type type)
         {
             if (_getDrawerTypeForPropertyAndTypeMethodInfo == null)
                 _getDrawerTypeForPropertyAndTypeMethodInfo = _scriptAttributeUtilityType.GetMethod("GetDrawerTypeForPropertyAndType", BindingFlags.Static | BindingFlags.NonPublic);
             return (Type)_getDrawerTypeForPropertyAndTypeMethodInfo!.Invoke(null, new object[] { property, type });
+        }
+
+        public static PropertyHandlerExtension GetHandler(SerializedProperty property)
+        {
+            if (_getHandlerMethodInfo == null)
+                _getHandlerMethodInfo = _scriptAttributeUtilityType.GetMethod("GetHandler", BindingFlags.Static | BindingFlags.NonPublic);
+            return new PropertyHandlerExtension(_getHandlerMethodInfo!.Invoke(null, new object[] { property }));
         }
     }
 }